task-tracker/requirements.txt
Flask==2.3.2
gspread==5.12.0
google-auth==2.28.1
oauth2client==4.1.3
python-dotenv==1.0.1
